# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TripSync API",
    "version": "1.0.0",
    "description": "API endpoints for TripSync - Trip Expense Splitter & Travel Planner backend."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["full_name", "username", "email", "password"],
        "properties": {
          "full_name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "username": { "type": "string", "pattern": "^[a-zA-Z0-9_]+$", "minLength": 2, "maxLength": 30 },
          "email": { "type": "string", "format": "email", "maxLength": 200 },
          "password": { "type": "string", "minLength": 8, "maxLength": 128 },
          "dob": { "type": "string", "format": "date", "nullable": true },
          "role": { "type": "string", "default": "user" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "user": { "type": "object" },
              "access_token": { "type": "string" },
              "refresh_token": { "type": "string" }
            }
          },
          "message": { "type": "string" }
        }
      },
...
      "CreateExpenseRequest": {
        "type": "object",
        "required": ["title", "amount", "date", "paid_by", "split_among"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "amount": { "type": "integer", "minimum": 1, "description": "Amount in paise (e.g. 100000 for 1000.00 INR)" },
          "currency": { "type": "string", "maxLength": 3, "default": "INR" },
          "category": { "type": "string", "enum": ["food", "transport", "accommodation", "activity", "shopping", "medical", "other"], "default": "other" },
          "date": { "type": "string", "format": "date" },
          "paid_by": { "type": "string", "description": "User ID of the member who paid" },
          "split_type": { "type": "string", "enum": ["equal", "custom"], "default": "equal" },
          "split_among": { "type": "array", "items": { "type": "string" } },
          "splits": { "type": "array", "items": { "type": "object" }, "description": "Used only for custom splits" },
          "notes": { "type": "string", "maxLength": 1000 },
          "receipts": { "type": "array", "items": { "type": "object" } }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created successfully" },
          "400": { "description": "Invalid payload/validation error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": { "description": "Unauthorized/Invalid credentials" }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Access token refreshed" },
          "401": { "description": "Invalid/Expired refresh token" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "summary": "Verify email OTP",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "otp": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP verified successfully" },
          "400": { "description": "Invalid/Expired OTP" }
        }
      }
    },
    "/auth/resend-otp": {
      "post": {
        "summary": "Resend email verification OTP",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP resent" }
        }
      }
    },
...
    "/trips/{trip_id}/expenses": {
      "post": {
        "summary": "Add a new expense",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Expense created" }
        }
      },
      "get": {
        "summary": "Get all expenses of a trip",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "paid_by", "in": "query", "schema": { "type": "string" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": { "description": "List of expenses" }
        }
      }
    },
    "/trips/{trip_id}/expenses/{expense_id}": {
      "get": {
        "summary": "Get an expense by ID",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "expense_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Expense details" }
        }
      },
      "delete": {
        "summary": "Delete an expense",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "expense_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Expense deleted" }
        }
      }
    },
...
  }
}